refactor: extract foreground anr detection logic to separate class
