+++ sdk/src/main/java/com/bugsnag/android/BlockedThreadDetector.java
@@ -1,15 +1,9 @@
 package com.bugsnag.android;
 
-import android.app.ActivityManager;
-import android.os.Build;
 import android.os.Handler;
 import android.os.HandlerThread;
 import android.os.Looper;
-import android.os.Process;
 import android.os.SystemClock;
-import android.support.annotation.Nullable;
-
-import java.util.List;
 
 /**
  * Detects whether a given thread is blocked by continuously posting a {@link Runnable} to it
@@ -38,22 +32,22 @@ final class BlockedThreadDetector {
     final Handler watchdogHandler;
     private final HandlerThread watchdogHandlerThread;
     final Delegate delegate;
-    final ActivityManager activityManager;
+    final ForegroundDetector foregroundDetector;
 
     volatile long lastUpdateMs;
     volatile boolean isAlreadyBlocked = false;
 
     BlockedThreadDetector(long blockedThresholdMs,
                           Looper looper,
-                          ActivityManager activityManager,
+                          ForegroundDetector foregroundDetector,
                           Delegate delegate) {
-        this(blockedThresholdMs, MIN_CHECK_INTERVAL_MS, looper, activityManager, delegate);
+        this(blockedThresholdMs, MIN_CHECK_INTERVAL_MS, looper, foregroundDetector, delegate);
     }
 
     BlockedThreadDetector(long blockedThresholdMs,
                           long checkIntervalMs,
                           Looper looper,
-                          ActivityManager activityManager,
+                          ForegroundDetector foregroundDetector,
                           Delegate delegate) {
         if ((blockedThresholdMs <= 0 || checkIntervalMs <= 0
             || looper == null || delegate == null)) {
@@ -64,7 +58,7 @@ final class BlockedThreadDetector {
         this.looper = looper;
         this.delegate = delegate;
         this.uiHandler = new Handler(looper);
-        this.activityManager = activityManager;
+        this.foregroundDetector = foregroundDetector;
 
         watchdogHandlerThread = new HandlerThread("bugsnag-anr-watchdog");
         watchdogHandlerThread.start();
@@ -104,7 +98,7 @@ final class BlockedThreadDetector {
 
     void checkIfThreadBlocked() {
         long delta = SystemClock.uptimeMillis() - lastUpdateMs;
-        boolean inForeground = isInForeground();
+        boolean inForeground = foregroundDetector.isInForeground();
 
         if (inForeground && delta > blockedThresholdMs) {
             if (!isAlreadyBlocked) {
@@ -114,48 +108,5 @@ final class BlockedThreadDetector {
         } else {
             isAlreadyBlocked = false;
         }
-    }
-
-    private boolean isInForeground() {
-        ActivityManager.RunningAppProcessInfo info = getProcessInfo();
-
-        if (info != null) {
-            return info.importance <= ActivityManager.RunningAppProcessInfo.IMPORTANCE_VISIBLE;
-        } else { // prefer a potential false positive if process info not available
-            return true;
-        }
-    }
-
-    private ActivityManager.RunningAppProcessInfo getProcessInfo() {
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            ActivityManager.RunningAppProcessInfo info =
-                new ActivityManager.RunningAppProcessInfo();
-            ActivityManager.getMyMemoryState(info);
-            return info;
-        } else {
-            return getProcessInfoPreApi16();
-        }
-    }
-
-    @Nullable
-    private ActivityManager.RunningAppProcessInfo getProcessInfoPreApi16() {
-        List<ActivityManager.RunningAppProcessInfo> appProcesses;
-
-        try {
-            appProcesses = activityManager.getRunningAppProcesses();
-        } catch (SecurityException exc) {
-            return null;
-        }
-
-        if (appProcesses != null) {
-            int pid = Process.myPid();
-
-            for (ActivityManager.RunningAppProcessInfo appProcess : appProcesses) {
-                if (pid == appProcess.pid) {
-                    return appProcess;
-                }
-            }
-        }
-        return null;
     }
 }
